requirements.txt
aiohttp==3.7.3
aiohttp-devtools==0.13.1
detect-secrets==0.14.3
GitPython==3.1.11
PyGithub==1.54
